+++ CHANGELOG.md
@@ -42,6 +42,19 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## [Unreleased]
 
+### Features
+
+- [#189] Add the flag `--valset-relay-mode` which allows a finer control over
+  how valsets will be relayed.
+
+### Deprecated
+
+- [#189] Deprecate the `--relay-valsets` flag.
+
+### Bug Fixes
+
+- [#189] Order validator set before deploying (`peggo bridge deploy-gravity`)
+
 ## [v0.2.4](https://github.com/umee-network/peggo/releases/tag/v0.2.4) - 2022-02-16
 
 ### Improvements